% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/CrossVal.R
\name{CrossVal}
\alias{CrossVal}
\title{Cross Validation}
\usage{
CrossVal(ART = NULL, S = NULL, phi = NULL, RiskF = NULL)
}
\arguments{
\item{ART}{Treatment status.}

\item{S}{Risk score.}

\item{RiskF}{Risk function.}

\item{Phi}{Percentage of taking viral load test.}
}
\value{
Cross-validation results.
}
\description{
This function allows you to compute summary statistics of
results from K-fold cross-validation.
}
\examples{
CrossVal(ART=NULL,S=NULL,phi=NULL,RiskF=NULL)
}
\keyword{Cross}
\keyword{validation.}

