% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/CV.TVLT.R
\name{CV.TVLT}
\alias{CV.TVLT}
\title{Cross Validation}
\usage{
CV.TVLT(Z, S, phi, K = 10, method = "nonpar", lambda)
}
\arguments{
\item{Z}{True disease status (No disease / treatment success coded as Z=0, diseased / treatment failure coded as Z=1).}

\item{S}{Risk score.}

\item{phi}{Percentage of patients taking viral load test.}

\item{K}{Number of folds in cross validation. The default is 10.}

\item{method}{The method to be used. The default is "nonpar", which returns result using nonparametric method. Another possible value is "semipar", which returns result estimated by semiparametric method assuming exponential tilt.}

\item{lambda}{A user-specified weight that reflects relative loss for the two types of misdiagnoses, taking value in \eqn{[0,1]}. \eqn{Loss=\lambda*I(FN)+(1-\lambda)*I(FP)}.}
}
\value{

Cross validated results on false classification rates (FNR, FPR), $\lambda-$ risk, total misclassification rate and AUC.

}
\description{
This function allows you to compute the average of misdiagnoses rate for viral failure and the optimal risk under min-\eqn{\lambda} rules
from K-fold cross-validation.
}
\examples{
d = Simdata
Z = d$Z # True Disease Status
S = d$S # Risk Score
phi = 0.1 #10\% of patients taking viral load test
lambda =0.5
CV.TVLT(Z, S, phi, K = 10, method = "semipar", lambda)
}
\keyword{Cross}
\keyword{FNR,}
\keyword{FPR.}
\keyword{optimal}
\keyword{risk,}
\keyword{validation,}

