% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Explore.R
\name{Explore}
\alias{Explore}
\title{Exploratory Analysis}
\usage{
Explore(ART, S = NULL, phi = NULL, RiskF = NULL)
}
\arguments{
\item{ART}{Treatment status.}

\item{S}{Risk score.}

\item{RiskF}{Risk function.}

\item{Phi}{Percentage of taking viral load test.}
}
\value{
Percentage of treatment failure (ART=1); histogram and summary of risk score by ART status.
}
\description{
This function allows you to compute the percentage of treatment failure (ART=1) and show distribution summary of risk score by ART status.
}
\examples{
Explore(ART,S=NULL,phi=NULL,RiskF=NULL)
}
\keyword{Exploratory}
\keyword{analysis,}
\keyword{distribution.}
\keyword{risk}
\keyword{score}

