% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Opt.semipar.rule.R
\name{Opt.semipar.rule}
\alias{Opt.semipar.rule}
\title{Optimal Semiparametric Rule}
\usage{
Opt.semipar.rule(Z, S, phi, lambda)
}
\arguments{
\item{Z}{True disease status (No disease / treatment success coded as Z=0, diseased / treatment failure coded as Z=1).}

\item{S}{Risk score.}

\item{phi}{Percentage of patients taking viral load test.}

\item{lambda}{A user-specified weight that reflects relative loss for the two types of misdiagnoses, taking value in \eqn{[0,1]}. \eqn{Loss=\lambda*I(FN)+(1-\lambda)*I(FP)}.}
}
\value{
Optimal semiparametric rule and its associated risk.
}
\description{
This function gives you the optimal semiparametric tripartite rule and the related optimal risk under min-\eqn{\lambda} rules.
}
\examples{
data = Simdata
Z = d$Z # True Disease Status
S = d$S # Risk Score
phi = 0.1 #10\% of patients taking viral load test
lambda = 0.5
Opt.semipar.rule( Z, S, phi, lambda)
}
\keyword{Semiparametric,}
\keyword{optimal}
\keyword{risk.}
\keyword{rules,}
\keyword{tripartite}

