% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClass.R
\docType{class}
\name{Output-class}
\alias{Output-class}
\title{Constructor of Output class}
\description{
This class contains invisible results of OptimalRule function.
}
\details{
\describe{
  \item{phi}{Percentage of patients taking viral load test.}
  \item{Z}{A vector of true disease status (Viral failure coded as Z=1).}
  \item{S}{A vector of risk Score.}
  \item{Rules}{A matrix of all possible tripartite rules (two cutoffs) derived from the training data set.}
  \item{Nonparametric}{A boolean indicating if nonparametric approach should be used in calculating the misclassfication rates. If FALSE, semiparametric approach would be used.}
  \item{FNR.FPR}{A matrix with two columns of misclassification rates, FNR and FPR.}
  \item{OptRule}{A numeric vector with two elements, the lower and upper cutoffs of the optimal tripartite rule.}
}


The Output class adds optimal rule to the TGST class.
}
