% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClass.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,}
\alias{TGST-method,}
\alias{Output-method}
\title{plot function.}
\usage{
\S4method{plot}{TGST}(x, y, ...)
}
\arguments{
\item{object}{Output object from \code{\link{TGST}} or \code{\link{Output}}.}
}
\value{
Distribution plot.
}
\description{
This function This function gives visualize object of class \code{TGST} or \code{Output}.
}
