% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ROC.semipar.R
\name{ROC.semipar}
\alias{ROC.semipar}
\title{Semiparametric ROC Analysis}
\usage{
ROC.semipar(Z, S, phi, plot = TRUE)
}
\arguments{
\item{Z}{True disease status (No disease / treatment success coded as Z=0, diseased / treatment failure coded as Z=1).}

\item{S}{Risk score.}

\item{phi}{Percentage of patients taking viral load test.}
}
\value{
AUC The area under the ROC curve.
Misdiagnoses rate for viral failure (i.e., false negative rate, FNR) and otherwise (i.e., false positive rate, FPR).
}
\description{
This function performs ROC analysis on the rules from nonparametric approach. If \eqn{plot=TRUE}, the ROC curve is returned.
}
\examples{
data = Simdata
Z = d$Z # True Disease Status
S = d$S # Risk Score
phi = 0.1 #10\% of patients taking viral load test
a = ROC.semipar( Z, S, phi,plot=TRUE)
a$AUC
a$FNR
a$FPR
}
\keyword{AUC,}
\keyword{FNR,}
\keyword{FPR.}
\keyword{Nonparametric,}
\keyword{ROC,}

