% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ROCAnalysis.R
\name{ROCAnalysis}
\alias{ROCAnalysis}
\title{ROC Analysis}
\usage{
ROCAnalysis(ART = NULL, S = NULL, phi = NULL, RiskF = NULL)
}
\arguments{
\item{ART}{Treatment status.}

\item{S}{Risk score.}

\item{RiskF}{Risk function.}

\item{Phi}{Percentage of taking viral load test.}
}
\value{
AUC, ROC curve.
}
\description{
This function allows you to do ROC analysis.
}
\examples{
ROCAnalysis(ART=NULL,S=NULL,phi=NULL,RiskF=NULL)
}
\keyword{ROC.}

