% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Rules.set.R
\name{Rules.set}
\alias{Rules.set}
\title{Nonparametric Rules}
\usage{
Rules.set(Z, S, phi)
}
\arguments{
\item{Z}{True disease status (No disease / treatment success coded as Z=0, diseased / treatment failure coded as Z=1).}

\item{S}{Risk score.}

\item{phi}{Percentage of patients taking viral load test.}
}
\value{
Cutoff points l and u;
Misdiagnoses rate for viral failure (i.e., false negative rate, FNR) and otherwise (i.e., false positive rate, FPR).
}
\description{
This function gives you all possible cutoffs \eqn{[l,u]} for tripartite rules, by applying nonparametric search to the given data.
\deqn{P(S in [l,u]) \le \phi}
}
\examples{
data = MiriamData
Z = (data$PVL_Count_Now>1000) #True Status
S = data$CD4_count_Now #Risk Score
phi = 0.1 #10\% of patients taking viral load test
Rules.set( Z, S, phi)
}
\keyword{FNR,}
\keyword{FPR.}
\keyword{Nonparametric,}
\keyword{rules,}
\keyword{tripartite}

