% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/SemiparaRule.R
\name{SemiparaRule}
\alias{SemiparaRule}
\title{Semiparametric Rules}
\usage{
SemiparaRule(ART = NULL, S = NULL, phi = NULL, RiskF = NULL)
}
\arguments{
\item{ART}{Treatment status.}

\item{S}{Risk score.}

\item{RiskF}{Risk function.}

\item{Phi}{Percentage of taking viral load test.}
}
\value{
Test.
}
\description{
This function allows you to do targeted viral load test under semiparametric rules.
}
\examples{
SemiparaRule(ART=NULL,S=NULL,phi=NULL,RiskF=NULL)
}
\keyword{Semiparametric,}
\keyword{decision}
\keyword{rule.}

