% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/DefineClass.R
\docType{methods}
\name{summary}
\alias{TVLT-method}
\alias{summary}
\alias{summary,}
\title{summary function.}
\usage{
\S4method{summary}{TVLT}(object, ...)
}
\arguments{
\item{object}{Output object from \code{\link{TVLT}}.}
}
\value{
Percentage of treatment failure;
Summary statistics (mean, standard deviation, minimum, median, maximum and IQR) of risk score by true disease status;
Distribution plot.
}
\description{
This function This function gives the summary of the data from \code{TVLT}.
}

