% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/DefineClass.R
\docType{class}
\name{TVLT-class}
\alias{TVLT-class}
\title{Constructor of TVLT class}
\description{
This class contains results of a run to search for all possible rules.
}
\details{
\describe{
  \item{phi}{Percentage of patients taking viral load test.}
  \item{Z}{A vector of true disease status (Viral failure coded as Z=1).}
  \item{S}{A vector of risk Score.}
  \item{Rules}{A matrix of all possible tripartite rules (two cutoffs) derived from the training data set.}
  \item{Nonparametric}{A boolean indicating if nonparametric approach should be used in calculating the misclassfication rates. If FALSE, semiparametric approach would be used.}
  \item{FNR.FPR}{A matrix with two columns of misclassification rates, FNR and FPR.}
}

If res is the result of rankclust(), each slot of results can be reached by res[k]@slotname, where
k is the number of clusters and slotname is the name of the slot we want to reach (see \link{Output-class}).
For the slots ll, bic, icl, res["slotname"] returns a vector of size K containing the values of the
slot for each number of clusters.
}

