% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClass.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,}
\alias{TGST-method}
\title{plot function.}
\usage{
plot(object)
}
\arguments{
\item{object}{Output object from \code{\link{TGST}}.}
}
\value{
Distribution plot.
}
\description{
This function gives the summary plot of the data from \code{TGST}.
}
