% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClass.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,TGST-method}
\title{summary function.}
\usage{
\S4method{summary}{TGST}(object)
}
\arguments{
\item{object}{Output object from \code{\link{TGST}}.}
}
\value{
Percentage of treatment failure; 
Summary statistics (mean, standard deviation, minimum, median, maximum and IQR) of risk score by true disease status; 
Distribution plot.
}
\description{
This function gives the summary of the data from \code{TGST}.
}
