% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/TVLT.summ.R
\name{TVLT.summ}
\alias{TVLT.summ}
\title{Summary of Disease Status and Risk Score}
\usage{
TVLT.summ(Z, S)
}
\arguments{
\item{Z}{True disease status (No disease / treatment success coded as Z=0, diseased / treatment failure coded as Z=1).}

\item{S}{Risk score.}
}
\value{
Percentage of treatment failure;
Summary statistics (mean, standard deviation, minimum, median, maximum and IQR) of risk score by true disease status;
Distribution plot.
}
\description{
This function allows you to compute the percentage of diseased (equivalent to treatment failure Z=1) and show distribution summary of risk score (S) by the true disease status (Z).
}
\examples{
data = read.csv("data.csv", header = TRUE)
Z = (data$PVL_Count_Now>1000) #True Status
S = data$CD4_count_Now #Risk Score
TVLT.summ(Z,S)
}
\keyword{(treatment}
\keyword{Prevalence}
\keyword{disease}
\keyword{distribution.}
\keyword{failure),}
\keyword{of}
\keyword{risk}
\keyword{score}
\keyword{summary,}

